Create INS1027Garratts.xml
According to my suggestion in https://github.com/BetaMasaheft/Documentation/issues/3158.

When creating the record, there was type="wd:" prefilled. There is no wikidata entry for Garratts Hall Library at the moment. I will try to login wikidata and create a record for Garratts Hall Library

I would like to add the information, that this was the private residence of the Lambert family, but I saw no space in the record for it.